Update readme, structure
