+++ web/html/src/eslint-local-rules/README.md
@@ -1,5 +1,8 @@
 # Local ESLint rules
 
-TODO: Write
+This directory contains local linter rules to enforce project-specific rules. This is a great option if you want to discourage or forbid certain types of practices.
 
-See https://github.com/cletusw/eslint-plugin-local-rules for more information
+Recommended reading:  
+
+ - https://eslint.org/docs/developer-guide/working-with-rules
+ - https://github.com/cletusw/eslint-plugin-local-rules
